Updated Error Handeling

+++ 18_Capstone/TenmoServer/DAO/AccountSqlDAO.cs
@@ -143,7 +143,7 @@ namespace TenmoServer.DAO
             {
                 AccountId = Convert.ToInt32(reader["account_id"]),
                 UserId = Convert.ToInt32(reader["user_id"]),
-                Balance = Convert.ToInt32(reader["balance"])
+                Balance = Convert.ToDecimal(reader["balance"])
             };
 
             return acc;
